+++ Cargo.toml
@@ -31,6 +31,7 @@ tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
 walkdir = "2.3.3"
 webbrowser = "0.8.9"
+youtube_dl = { version = "0.8.0", features = ["tokio", "downloader"] }
 
 [target.'cfg(unix)'.dependencies]
 tar = "0.4"
