+++ src/server/handlers/mod.rs
@@ -1,0 +1,10 @@
+use crate::{data::database::Database, service::generator::CompilationGenerator};
+
+pub mod common;
+pub mod local;
+pub mod stash;
+
+pub struct AppState {
+    pub generator: CompilationGenerator,
+    pub database: Database,
+}
